Optimizacion NavBar
