+++ src/main/java/org/proyecto/retroshare/controller/GameConsoleRestController.java
@@ -18,7 +18,9 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.multipart.MultipartFile;
 
 @CrossOrigin("*")
 @RestController
@@ -57,6 +59,7 @@ public class GameConsoleRestController {
 		user.getProductOwner().add(gameConsole);
 
 		ProductStatus productStatus = productStatusRepository.getOne(idProductStatus);
+		
 		gameConsole.setProductStatus(productStatus);
 		productStatus.getProducts().add(gameConsole);
